chore: move cli related funcs to own module
