+++ dexios/src/subcommands/decrypt.rs
@@ -1,7 +1,7 @@
 use std::process::exit;
 use std::sync::Arc;
 
-use super::prompt::overwrite_check;
+use crate::cli::prompt::overwrite_check;
 use crate::global::states::{EraseMode, HashMode, HeaderLocation, PasswordState};
 use crate::global::structs::CryptoParams;
